backend/requirements.txt
Django==2.2.12
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
django-filter==2.2.0
django-cors-headers==3.2.1
django-import-export==2.0.2
ansible==2.7.10
redis==3.4.1
celery==4.4.2
django-celery-results
django-celery-beat
drf-extensions
eventlet
mysqlclient
markdown==3.2.1
pillow==7.1.2
simpleui
coreapi
channels
channels_redis
flower
django-mdeditor
markdown
Pygments
